+++ CookieeApp/app/index.js
@@ -1,17 +1,21 @@
 import React, { useEffect, useState } from "react";
-import { View, StyleSheet, Text, Image } from "react-native";
+import { View, StyleSheet, Text, Image, Pressable } from "react-native";
 import AppLoading from "expo-app-loading";
 import { StatusBar } from "expo-status-bar";
 
-import Login from "./screens/Login";
-import Loading from "./screens/Loading";
-import User from "./screens/User";
-import CalendarHome from "./screens/CalenderHome";
+import Login from "./screen/Login";
+import Loading from "./screen/Loading";
+import User from "./screen/User";
+// import CalendarHome from "./screens/CalenderHome";
 
 import "expo-router/entry";
-import { useRouter } from "expo-router";
+import { Link, router } from "expo-router";
 
 export default function App() {
+  const handlePress = () => {
+    router.push("/screen/CalenderHome");
+  };
+
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [ready, setReady] = useState(true);
 
@@ -36,7 +40,7 @@ export default function App() {
   }
 
   return (
-    <View style={styles.container}>
+    <View>
       {/* {isLoggedIn ? (
         // 사용자가 로그인한 경우 메인 화면
         <>
@@ -48,7 +52,19 @@ export default function App() {
         <Login onLogin={handleLogin} />
       )}
       <StatusBar style="auto" /> */}
-      <CalendarHome />
+      {/* <CalendarHome /> */}
+
+      <Pressable onPress={handlePress}>
+        <Text>00000</Text>
+      </Pressable>
+
+      <Link href={"/screen/CalenderHome"} asChild>
+        <Pressable>
+          <Text style={{ fontSize: 20, color: "black" }}>
+            "index" to "/screen/CalenderHome"
+          </Text>
+        </Pressable>
+      </Link>
     </View>
   );
 }
@@ -58,5 +74,6 @@ const styles = StyleSheet.create({
     alignItems: "center",
     justifyContent: "center",
     flex: 1,
+    backgroundColor: "#000000",
   },
 });
